feature: optimize logic in retries

+++ remove.py
@@ -7,6 +7,8 @@ def remove(file='verify.out'):
         with open(file, 'r') as f:
             packages = list(map(lambda line: str(line).strip(), f.readlines()))
             for package in packages:
+                if package == '':
+                    continue
                 path = 'packages/{}'.format(package)
                 if os.path.exists(path):
                     shutil.rmtree(path)
